requirements.txt
absl-py==0.13.0
anyio==3.3.0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
configparser==5.0.2
cryptography==3.4.7
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-appconf==1.0.4
django-bootstrap3==15.0.0
django-crispy-forms==1.12.0
django-embed-video==1.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-password-reset==2.0
entrypoints==0.3
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jax==0.2.17
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.1
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
opt-einsum==3.3.0
package-name==0.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2>=2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pyzmq==22.1.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0


gunicorn==20.1.0
static3==0.7.0
dj-database-url==0.5.0
dj-static==0.0.6

